Modify vulture configuration
To apply the vulture allow list, vulture must be invoked like
`vulture TARGET ALLOW_LIST_FILE`. The allow list is not configurable!
This change is a hack, as pointing vulture at . will then automatically
include the allow list.

This allows devs to simply invoke `vulture` and not concern themselves
with specifying the allow list.