Symlink templates folder for each project template
Apparently cookiecutter template inheritance is bounded by the project
template directory. That is, when using nested project templates, a `templates`
folder cannot be shared amongst the project templates because it only searches
the project template's directory, not the cookiecutter project as a whole. In
order to get around this limitation, I added symlinks to the templates
directory in each of the project template directories.